전체적인 파일 구조 정리
