+++ 7th-be-blog/build.gradle
@@ -11,12 +11,13 @@ java {
     sourceCompatibility = '21'
 }
 
-repositories {
+repositories {1
     mavenCentral()
 }
 
 dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-web'
+    implementation 'org.springframework.boot:spring-boot-starter-validation'
     implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.5.0'
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
