--- 7th-be-blog/src/main/java/com/leets/blog/common/ApiResponse.java
@@ -1,31 +1,0 @@
-package com.leets.blog.common;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-@JsonPropertyOrder({"isSuccess", "code", "message", "result"})
-public class ApiResponse<T> {
-
-    @JsonProperty("isSuccess")
-    private final Boolean isSuccess;
-    private final String code;
-    private final String message;
-    
-    @JsonInclude(JsonInclude.Include.NON_NULL)
-    private final T result;
-
-    // 성공 시
-    public static <T> ApiResponse<T> onSuccess(T result) {
-        return new ApiResponse<>(true, "COMMON200", "요청에 성공하였습니다.", result);
-    }
-
-    // 실패 시 (상황에 따라 result에 에러 정보를 담거나 null 처리)
-    public static <T> ApiResponse<T> onFailure(String code, String message, T result) {
-        return new ApiResponse<>(false, code, message, result);
-    }
-}
