--- 7th-be-blog/src/main/java/com/leets/blog/common/BaseErrorCode.java
@@ -1,22 +1,0 @@
-package com.leets.blog.common;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-@Getter
-@AllArgsConstructor
-public enum BaseErrorCode {
-    // 공통 에러
-    SUCCESS(HttpStatus.OK, "COMMON200", "요청에 성공하였습니다."),
-    BAD_REQUEST(HttpStatus.BAD_REQUEST, "COMMON400", "잘못된 요청입니다."),
-    INTERNAL_SERVER_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, "COMMON500", "서버 내부 오류가 발생했습니다."),
-    
-    // 도메인별 에러 (예시)
-    USER_NOT_FOUND(HttpStatus.NOT_FOUND, "USER4001", "사용자를 찾을 수 없습니다."),
-    POST_NOT_FOUND(HttpStatus.NOT_FOUND, "POST4001", "게시글을 찾을 수 없습니다.");
-
-    private final HttpStatus httpStatus;
-    private final String code;
-    private final String message;
-}
