--- 7th-be-blog/src/main/java/com/leets/blog/common/ErrorResponse.java
@@ -1,31 +1,0 @@
-package com.leets.blog.common;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import java.time.LocalDateTime;
-
-@Getter
-@Builder
-@AllArgsConstructor
-public class ErrorResponse {
-    private final String code;
-    private final String message;
-    private final LocalDateTime timestamp;
-
-    public static ErrorResponse of(BaseErrorCode errorCode) {
-        return ErrorResponse.builder()
-                .code(errorCode.getCode())
-                .message(errorCode.getMessage())
-                .timestamp(LocalDateTime.now())
-                .build();
-    }
-
-    public static ErrorResponse of(String code, String message) {
-        return ErrorResponse.builder()
-                .code(code)
-                .message(message)
-                .timestamp(LocalDateTime.now())
-                .build();
-    }
-}
