--- 7th-be-blog/src/main/java/com/leets/blog/controller/HealthController.java
@@ -1,14 +1,0 @@
-package com.leets.blog.controller;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/health")
-public class HealthController {
-    @GetMapping
-    public String health() {
-        return "OK";
-    }
-}
