+++ 7th-be-blog/src/main/java/com/leets/blog/domain/health/controller/HealthController.java
@@ -1,0 +1,14 @@
+package com.leets.blog.domain.health.controller;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/health")
+public class HealthController {
+    @GetMapping
+    public String health() {
+        return "OK";
+    }
+}
