+++ 7th-be-blog/src/main/java/com/leets/blog/domain/string/dto/StringRequestDto.java
@@ -1,0 +1,12 @@
+package com.leets.blog.domain.string.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class StringRequestDto {
+    private String value;
+}
