+++ 7th-be-blog/src/main/java/com/leets/blog/domain/string/dto/StringResponseDto.java
@@ -1,0 +1,13 @@
+package com.leets.blog.domain.string.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class StringResponseDto {
+
+    private final String string_one;
+    private final String string_two;
+
+}
