+++ 7th-be-blog/src/main/java/com/leets/blog/domain/string/service/StringService.java
@@ -1,0 +1,10 @@
+package com.leets.blog.domain.string.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class StringService {
+    public String repeat(String value) {
+        return value;
+    }
+}
