+++ 7th-be-blog/src/main/java/com/leets/blog/domain/user/entity/User.java
@@ -1,0 +1,48 @@
+package com.leets.blog.domain.user.entity;
+
+import com.leets.blog.global.common.BaseEntity;
+import com.leets.blog.domain.post.entity.Post;
+import com.leets.blog.domain.comment.entity.Comment;
+import jakarta.persistence.*;
+import lombok.*;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+@Builder
+@Table(name = "users")
+public class User extends BaseEntity {
+
+    // - pk
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    // - 유저 이름
+    @Column(nullable = false)
+    private String name;
+
+    // - 이메일
+    @Column(nullable = false, unique = true)
+    private String email;
+
+    // - 패스워드
+    @Column(nullable = false)
+    private String password;
+
+    // user - post = 1 : N
+    @OneToMany(mappedBy = "user")
+    private List<Post> posts = new ArrayList<>();
+
+    // user - comment = 1 : N
+    @OneToMany(mappedBy = "user")
+    private List<Comment> comments = new ArrayList<>();
+
+    // user 데이터를 지우지 않고 상태변경 -> FK 보존
+    @Column(name = "is_deleted", nullable = false)
+    private boolean isDeleted = false;
+}
