+++ 7th-be-blog/src/main/java/com/leets/blog/domain/user/repository/UserRepository.java
@@ -1,0 +1,9 @@
+package com.leets.blog.domain.user.repository;
+
+import com.leets.blog.domain.user.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+}
