--- 7th-be-blog/src/main/java/com/leets/blog/dto/StringRequestDto.java
@@ -1,12 +1,0 @@
-package com.leets.blog.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class StringRequestDto {
-    private String value;
-}
