--- 7th-be-blog/src/main/java/com/leets/blog/dto/StringResponseDto.java
@@ -1,13 +1,0 @@
-package com.leets.blog.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class StringResponseDto {
-
-    private final String string_one;
-    private final String string_two;
-
-}
