--- 7th-be-blog/src/main/java/com/leets/blog/dto/post/PostRequest.java
@@ -1,24 +1,0 @@
-package com.leets.blog.dto.post;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Size;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class PostRequest {
-    
-    @NotBlank(message = "제목은 필수입니다.")
-    @Size(max = 100, message = "제목은 100자 이내여야 합니다.")
-    private String title;
-
-    @NotBlank(message = "내용은 필수입니다.")
-    private String content;
-
-    @NotNull(message = "작성자 ID는 필수입니다.")
-    private Long userId;
-}
