--- 7th-be-blog/src/main/java/com/leets/blog/dto/post/PostResponse.java
@@ -1,17 +1,0 @@
-package com.leets.blog.dto.post;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import java.time.LocalDateTime;
-
-@Getter
-@Builder
-@AllArgsConstructor
-public class PostResponse {
-    private Long id;
-    private String title;
-    private String content;
-    private String authorName;
-    private LocalDateTime createdAt;
-}
