+++ 7th-be-blog/src/main/java/com/leets/blog/global/exception/PostException.java
@@ -1,0 +1,13 @@
+package com.leets.blog.global.exception;
+
+import com.leets.blog.global.common.BaseErrorCode;
+
+public class PostException extends GeneralException {
+    public PostException(BaseErrorCode errorCode) {
+        super(errorCode);
+    }
+
+    public PostException(BaseErrorCode errorCode, String message) {
+        super(errorCode, message);
+    }
+}
