--- 7th-be-blog/src/main/java/com/leets/blog/repository/UserRepository.java
@@ -1,9 +1,0 @@
-package com.leets.blog.repository;
-
-import com.leets.blog.entity.user.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-}
