--- 7th-be-blog/src/main/java/com/leets/blog/service/StringService.java
@@ -1,10 +1,0 @@
-package com.leets.blog.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class StringService {
-    public String repeat(String value) {
-        return value;
-    }
-}
